requirements.txt
PyMuPDF
openpyxl
Pillow
python-docx
lxml
